# [FLINK-10727][network] remove unnecessary synchronization in SingleInputGate#requestPartitions()