@NicoK 2018-10-31
Thanks @zhijiangW and @pnowojski for the reviews.
merging...