@tillrohrmann 2018-11-02
I think making `requestedPartitionsFlag` volatile won't solve the problem.

Let's continue the discussion on the PR thread of #6992.